Move server tasks to their own file.
